requirements.txt
streamlit~=1.39.0
neo4j~=5.26.0
requests~=2.32.0

python-dotenv~=0.21.0